Check for existence of certain directories
No need to clone repos that already exist locally

+++ generic/zsh-customizations.sh
@@ -1,23 +1,33 @@
 #!/usr/bin/env bash
-
-sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [[ ! -d $HOME/.oh-my-zsh ]]
+then
+    sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+fi
 
 # plugins
 # Autosuggestions
-git clone --depth=1 https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-rm -rf  ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions/.git
-rm -rf  ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions/.github
-rm -rf  ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions/.circleci
+if [[ ! -d $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions ]]
+then
+    git clone --depth=1 https://github.com/zsh-users/zsh-autosuggestions $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+    rm -rf  $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions/.git
+    rm -rf  $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions/.github
+    rm -rf  $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions/.circleci
+fi
 
 # Syntax Highlighting
-git clone --depth=1 https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
-rm -rf ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting/.git
-
+if [[ ! -d $HOME/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting ]]
+then
+    git clone --depth=1 https://github.com/zsh-users/zsh-syntax-highlighting.git $HOME/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting
+    rm -rf $HOME/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting/.git
+fi
 
 # Pure prompt
-mkdir -p "$HOME/.zsh"
-git clone --depth=1 https://github.com/sindresorhus/pure.git "$HOME/.zsh/pure"
-rm -rf $HOME/.zsh/pure/.git
+if [[ ! -d $HOME/.zsh/pure ]]
+then
+    mkdir -p "$HOME/.zsh"
+    git clone --depth=1 https://github.com/sindresorhus/pure.git $HOME/.zsh/pure
+    rm -rf $HOME/.zsh/pure/.git
+fi
 
 exit
